requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
gunicorn==21.2.0
supabase==2.4.2
python-dotenv==1.0.0
pandas==2.2.0
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.6.4
pydantic-settings==2.1.0
slowapi==0.1.9
redis==5.0.1
